update bf16-true support

+++ nova/ebt/modeling_ebt.py
@@ -85,9 +85,9 @@ class EBT_NLP(LightningModule):
         if self.hparams.normalize_initial_condition:
             if self.hparams.normalize_initial_condition_only_first_step:
                 if mcmc_step == 0:
-                    predicted_tokens = self.softmax(predicted_tokens.float()).to(predicted_tokens.dtype)
+                    predicted_tokens = self.softmax(predicted_tokens)
             else:
-                predicted_tokens = self.softmax(predicted_tokens.float()).to(predicted_tokens.dtype)
+                predicted_tokens = self.softmax(predicted_tokens)
                 
             if self.hparams.vocab_to_embed_uses_prob_dist: # predicted_embeds is B, S, V; embed is V, D
                 predicted_embeddings = torch.matmul(predicted_tokens, self.embeddings.weight) #BS, S, D
@@ -296,7 +296,7 @@ class EBT_NLP(LightningModule):
         if self.hparams.denoising_initial_condition == "most_recent_embedding":
             raise NotImplementedError(f"most_recent_embedding denoising_initial_condition not supported for NLP yet")
         elif self.hparams.denoising_initial_condition == "random_noise":
-            predicted_tokens = torch.randn(size=(embeddings.shape[0], embeddings.shape[1], self.vocab_size), dtype=torch.bfloat16, device = self.device) * self.hparams.gaussian_random_noise_scaling
+            predicted_tokens = torch.randn(size=(embeddings.shape[0], embeddings.shape[1], self.vocab_size), dtype=embeddings.dtype, device=self.device) * self.hparams.gaussian_random_noise_scaling
         elif self.hparams.denoising_initial_condition == "zeros":
             predicted_tokens = torch.zeros(size=(embeddings.shape[0], embeddings.shape[1], self.vocab_size), device = self.device)
         else:
